Fix parallel build dependency, close #65
This commit corrects the dependency issue in the Makefile which was causing errors during parallel builds. Specifically, it addresses the following changes:

- Separate build rules for local/graftcp-local and local/mgraftcp to individually specify their dependencies and build commands.
- Ensure local/mgraftcp explicitly depends on libgraftcp.a, and build it prior to local/mgraftcp to resolve potential race conditions.
- The local/graftcp-local target does not depend on libgraftcp.a, hence a separate rule is created without this dependency.

These adjustments ensure that each binary is built correctly respecting their precise dependencies, and thus preventing errors that may occur due to incorrect builds when using the -j option for parallel execution in make.