Add support for running traced commands as a specified user
This change adds a new command-line option `-u/--user=<username>`.
The primary motivation for this feature is to address the limitations when the traced command invokes a `setuid` system call (such as those executed by the `yay` command, which internally calls `sudo pacman ...`). Ordinary users cannot `ptrace` a `setuid` call, which leads to errors like "sudo: effective uid is not 0..." and the subsequent termination of the process.

To circumvent this issue, the tracer process can start with root privileges, and the tracee process executes the commands to be traced as the specified user. For example:

    sudo mgraftcp -u username command_to_trace

This allows tracing commands that invoke setuid system calls, which would
otherwise fail due to insufficient privileges when run as a non-root user.